Revert "rename dir"
This reverts commit b214a74bf4c64cca6e71dafff24fe236e5ddf193.
