+++ test/smoke/src/areas/positron/console/consoleANSI.test.ts
@@ -7,7 +7,7 @@
 import { expect } from '@playwright/test';
 import { Application, PositronRFixtures } from '../../../../../automation';
 import { join } from 'path';
-import { setupAndStartApp } from '../../../test-runner/test-hooks';
+import { setupAndStartApp } from '../../../test-helpers/test-hooks';
 
 describe('Console ANSI styling', () => {
 	setupAndStartApp();
