+++ test/smoke/src/areas/positron/test-explorer/test-explorer.test.ts
@@ -6,7 +6,7 @@
 import path = require('path');
 import { Application, PositronRFixtures, PositronUserSettingsFixtures } from '../../../../../automation';
 import { expect } from '@playwright/test';
-import { setupAndStartApp } from '../../../test-runner/test-hooks';
+import { setupAndStartApp } from '../../../test-helpers/test-hooks';
 
 describe('Test Explorer', () => {
 	setupAndStartApp();
