+++ test/smoke/src/areas/positron/top-action-bar/top-action-bar.test.ts
@@ -6,7 +6,7 @@
 import { join } from 'path';
 import { expect } from '@playwright/test';
 import { Application, PositronUserSettingsFixtures } from '../../../../../automation';
-import { setupAndStartApp } from '../../../test-runner/test-hooks';
+import { setupAndStartApp } from '../../../test-helpers/test-hooks';
 
 describe('Top Action Bar #web', () => {
 	setupAndStartApp();
