+++ test/smoke/src/test-helpers/config.ts
