+++ test/smoke/src/test-helpers/testSetup.ts
@@ -7,7 +7,7 @@ import * as fs from 'fs';
 import { join } from 'path';
 import rimraf = require('rimraf');
 import mkdirp = require('mkdirp');
-import { getBuildElectronPath, getDevElectronPath, Logger } from '../../../automation/out';
+import { getBuildElectronPath, getDevElectronPath, Logger } from '../../../automation';
 import { createLogger } from './logger';
 
 const ROOT_PATH = process.env.ROOT_PATH || 'ROOT_PATH not set';
